requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
h5py
pydantic
python-multipart